backend/pyproject.toml
[project]
name = "codebase-analyzer"
version = "0.1.0"
description = "AI-Powered Codebase Analyzer — Portfolio Project"
requires-python = ">=3.11"

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",

    # Config & validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Database
    "sqlmodel>=0.0.19",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",

    # Git & file handling
    "gitpython>=3.1.43",
    "pathspec>=0.12.1",
    "chardet>=5.2.0",

    # AST parsing
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-go>=0.23.0",

    # Embeddings & vector store
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",

    # Retrieval
    "rank-bm25>=0.2.2",
    "flashrank>=0.2.0",
    "tiktoken>=0.7.0",

    # Dimensionality reduction
    "umap-learn>=0.5.6",
    "hdbscan>=0.8.33",
    "numpy>=1.26.0",

    # LLM
    "litellm>=1.40.0",
    "anthropic>=0.28.0",

    # Rate limiting
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]